lint + click outside hook
